Completely functional code
The main class is fully functional, additional code is present as this is used for my Thesis (Bulk solution, reactions, reaction names).
Lastly, some examples of code execution are included. For future use, the class could be transformed to a .py file to allow importing it and being easier on the eyes. The Gibbs energy file can be expanded, but a new instance of the class would need to be made after that change.